+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v1.1.0 (in development)
 
+- Use the `latest` Babel preset
 - Remove the use of `setTimeout` when processing subscribers
 
 # v1.0.1